+++ src/util.h
