+++ src/util_common.h
